+++ NEWS.md
@@ -1,3 +1,15 @@
+## 0.10.0 to 0.10.1
+
+ * ContactDB: Deleting organisations:
+   The comparison of the organisation data stored in the database and the
+   organisation data provided by the frontend now ignores empty 'expires'
+   entries in all annotation types.
+
+## 0.9.2 to 0.10.0
+
+ * For the Audit log `intelmq-certbund-contact` version 0.9.5 is
+   required for the `audit_log` database table.
+
 ## 0.9.1 to 0.9.2
  
  * Fixed compatibility of contactdb_api with PostsgreSQL >= 10.0.